Implement user-friendly names in URL using PathCache
This is kinda magic: we use the user-visible names (display names) in path
and once we get the URL with this patch, we try to map it back to GDrive
UIDs using the PatchCache and falling back to recursively resolving UIDs
via FileFetchJob's Search functinality.

Opening/listing still seems to be pretty fast, so yay for us!

The only thing that currently does not work is opening of files that we
have converted from Google Documents format, because we have manually
changed their extension (added .odt, .png , ...) so obviously the names no
longer map to anything on the server and so the FileFetchJob search fails
to resolve the UID. I'm not sure if there's any way to solve it other than
stop appending custom filename extensions to the converted files.