A little hack to uniquely identify files
For each files we append ?id=%GDRIVE_UID% query item to UDS_URL attribute.
Users still see the pretty filenames everywhere, but we can easily identify
the files now. We will still fallback to using PathCache and recursive server
queries in case the id query is missing - either we got some URL copied from
somewhere, or we are working with folders.

The reason this does not work with folders is that folder cannot have query
item, so KIO converts the URL to gdrive://acount/some/folder/?id=XYZ - it
creates an empty file/folder, which then breaks cdup (and some other things).