Add gdrive:/ link to remote:/
This replaces our custom Dolphin .desktop file (which was just a
temporary solution) with a link to the `gdrive:/` URL in `remote:/`.

This way KIO GDrive can be easily reached from the Dolphin's (or
Konqueror's) Network folders or from the Plasma's Folder View widget.

Currently this only works on systems with plasma-workspace installed.
Long-term plan is to move the remote:/ ioslave from plasma-workspace to
KIO.

Task: T4065