appstream: make kio-gdrive an addon component
Summary:
The `desktop` type is the deprecated version of the
`desktop-application` type and is reserved for desktop applications,
which are supposed to install a .desktop file.

kio-gdrive is not a desktop application, but rather an addon.

Since we don't have a generic KIO component that we can extend, for now
we just extend all the file managers that support KIO.

See https://lists.debian.org/debian-qt-kde/2018/07/msg00096.html for context.

Differential Revision: https://phabricator.kde.org/D15221