Introduce support for Shared With Me files
'Shared With Me' files don't have a folderId as parent, so in listDir()
we need to adapt the query.

In resolveFileIdFromPath() we also need to check whether we have a 'top-level'
shared-with-me file. In that case, we avoid the recursion and just skip the parentId in the query.

FEATURE: 385388
FIXED-IN: 22.12.0